// src/utils/type.ts
export type GroupFacebook = {
    id?: string;
    topic_id: string;
    group_id: string;
    name: string;
}
// ...
export type Topic = {
    id?: string;
    name: string;
    type: string;
    description: string;
    is_priority?: boolean;
};